deleted createFilePermissions function and created base code for updateFolderPermissions
